Melhorias no backup de servidores

+++ src/main/java/com/kuuhaku/model/persistent/Backup.java
@@ -103,15 +103,16 @@ public class Backup {
 		LinkedList<Long> oldIDs = new LinkedList<>();
 		LinkedList<GuildCategory> oldCategories = new LinkedList<>();
 
-		/*g.getChannels().forEach(chn -> {
+		g.getChannels().forEach(chn -> {
 			try {
 				queue.offer(new request(0, chn.delete()));
 			} catch (Exception ignore) {
 			}
 		});
+
 		g.getRoles().forEach(r -> {
 			try {
-				queue.offer(new request(0, r.delete()));
+				if (!r.isPublicRole()) queue.offer(new request(0, r.delete()));
 			} catch (Exception ignore) {
 			}
 		});
@@ -119,7 +120,7 @@ public class Backup {
 		gdata.getCategories().forEach(gc -> {
 			queue.offer(new request(1, g.createCategory(gc.getName())));
 			oldCategories.offer(gc);
-		});*/
+		});
 
 		gdata.getRoles().forEach(gr -> {
 			queue.offer(new request(2, g.createRole()
@@ -134,8 +135,7 @@ public class Backup {
 		Executors.newSingleThreadExecutor().execute(() -> {
 			while (!queue.isEmpty()) {
 				try {
-					queue.poll().roleAct.complete();
-					/*request act = queue.poll();
+					request act = queue.poll();
 					switch (act.type) {
 						case 0:
 							act.clearAct.complete();
@@ -146,7 +146,7 @@ public class Backup {
 						case 2:
 							newRoles.put(oldIDs.poll(), act.roleAct.complete());
 							break;
-					}*/
+					}
 
 					Thread.sleep(500);
 				} catch (InterruptedException e) {
